+++ src/components/index.ts
@@ -1,3 +1,4 @@
 import ResourceCard from "./resource-card/resource-card";
+import ContactForm from "./contact-form/contact-form";
 
-export { ResourceCard };
+export { ResourceCard, ContactForm };
